Closes trac:631: one-sided fixes
Submitted by brbarret, Reviewed by Galen, Merged by Tim Mattox

r12667:
 * Use one-sided datatype check instead of send/receive and check both
the origin and target datatypes 
 * Allow error handler to be set on MPI_WIN_NULL, per standard 
 * Allow recursive calls into the pt2pt osc component's progress function 
 * Fix an uninitialized variable problem in the unlock header 

r12673:
 * Fix a counter roll-over issue that could result from a large
(but not excessive) number of outstanding put/get/accumulate calls
during a single synchronization issues
 * Fix epoch issue with rdma component that would effect PWSC synchronization

r12675:
Remove check for epoch on the receiver side, as this is a race
condition between message arrival, and because we'll be still
blocking in a synchronization call, is of no consequence to the user.
We do an epoch check on the sending side, so this isn't an issue there either.

r12676:
Fix issue where messages could start arriving before the window was
fully initialized, especially during lock/unlock, which doesn't use MPI
collectives for synchronization (unlike Fence)

This commit was SVN r12681.

The following SVN revision numbers were found above:
  r12667 --> open-mpi/ompi@63e5668e2940cf3e98e5cfe447786776aae795ec
  r12673 --> open-mpi/ompi@0c25f7be092348ad8691ac93b774341810dc4252
  r12675 --> open-mpi/ompi@b5b6f4c4bfbd6b7c913467948d04ff1d7eea038a
  r12676 --> open-mpi/ompi@bc3250fe6f4eb29d1e7cf3ebf36f874508a106d6

The following Trac tickets were found above:
  Ticket 631 --> https://svn.open-mpi.org/trac/ompi/ticket/631

+++ ompi/mca/osc/rdma/osc_rdma.h
@@ -72,10 +72,10 @@ struct ompi_osc_rdma_module_t {
         started.  p2p_lock must be held when modifying this field. */
     opal_list_t p2p_pending_sendreqs;
 
-    /** list of int16_t counters for the number of requests to a
+    /** list of unsigned int counters for the number of requests to a
         particular rank in p2p_comm for this access epoc.  p2p_lock
         must be held when modifying this field */
-    short *p2p_num_pending_sendreqs;
+    unsigned int *p2p_num_pending_sendreqs;
 
     /** For MPI_Fence synchronization, the number of messages to send
         in epoch.  For Start/Complete, the number of updates for this
@@ -109,7 +109,7 @@ struct ompi_osc_rdma_module_t {
     opal_list_t p2p_long_msgs;
 
     opal_list_t p2p_copy_pending_sendreqs;
-    short *p2p_copy_num_pending_sendreqs;
+    unsigned int *p2p_copy_num_pending_sendreqs;
 
     bool p2p_eager_send;
 
@@ -117,9 +117,9 @@ struct ompi_osc_rdma_module_t {
     /* an array of <sizeof(p2p_comm)> ints, each containing the value
        1. */
     int *p2p_fence_coll_counts;
-    /* an array of <sizeof(p2p_comm)> shorts, for use in experimenting
+    /* an array of <sizeof(p2p_comm)> unsigned ints, for use in experimenting
        with different synchronization costs */
-    short *p2p_fence_coll_results;
+    unsigned int *p2p_fence_coll_results;
 
     mca_osc_fence_sync_t p2p_fence_sync_type;
